according to issue #217
transition log shall be debug instead of info.

TransitionLogger can be a class by itself
