+++ styx-common/src/main/java/com/spotify/styx/util/ReplayEvents.java
@@ -20,12 +20,9 @@
 
 package com.spotify.styx.util;
 
-import static java.lang.String.format;
-
 import com.google.common.base.Throwables;
 import com.spotify.styx.model.SequenceEvent;
 import com.spotify.styx.model.WorkflowInstance;
-import com.spotify.styx.state.OutputHandler;
 import com.spotify.styx.state.RunState;
 import com.spotify.styx.storage.Storage;
 import java.io.IOException;
@@ -33,17 +30,13 @@ import java.time.Instant;
 import java.util.Map;
 import java.util.Optional;
 import java.util.SortedSet;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public final class ReplayEvents {
 
-  private static final Logger LOG = LoggerFactory.getLogger(ReplayEvents.class);
-
   private ReplayEvents() {
+    throw new UnsupportedOperationException();
   }
 
-  // TODO: fix NPath complexity
   public static Optional<RunState> getBackfillRunState(
       WorkflowInstance workflowInstance,
       Map<WorkflowInstance, RunState> activeWorkflowInstances,
@@ -96,43 +89,6 @@ public final class ReplayEvents {
       }
     }
     return backfillFound ? Optional.of(restoredState) : Optional.empty();
-  }
-
-  public static OutputHandler transitionLogger(String prefix) {
-    return (state) -> {
-      final String instanceKey = state.workflowInstance().toKey();
-      LOG.info(
-          "{}{} transition -> {} {}",
-          prefix, instanceKey, state.state().name().toLowerCase(), stateInfo(state));
-    };
-  }
-
-  private static String stateInfo(RunState state) {
-    switch (state.state()) {
-      case NEW:
-      case PREPARE:
-      case ERROR:
-      case DONE:
-        return format("tries:%d", state.data().tries());
-
-      case SUBMITTED:
-      case RUNNING:
-      case FAILED:
-        return format("tries:%d execId:%s",
-            state.data().tries(), state.data().executionId());
-
-      case TERMINATED:
-        return format("tries:%d execId:%s exitCode:%s",
-            state.data().tries(), state.data().executionId(), state.data().lastExit().map(
-                String::valueOf).orElse("-"));
-
-      case QUEUED:
-        return format("tries:%d delayMs:%s",
-            state.data().tries(), state.data().retryDelayMillis());
-
-      default:
-        return "";
-    }
   }
 
   private static final class SettableTime implements Time {
